requirements/tests.txt
-r base.txt
pytest>=6
pytest-django
flake8
pytest-cov
model-bakery
django-cors-headers